requirements.txt
requests
beautifulsoup4
pillow
PySide6
pyinstaller
